ASN.1 encode. Adding in Bit String formatting rules
